requirements.txt
accelerate==1.0.1
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
diffusers==0.30.3
einops==0.8.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gpustat==1.1.1
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.2.1
numpy==1.24.0
nvidia-ml-py==12.560.30
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
psutil==6.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
tifffile==2024.8.30
tokenizers==0.20.1
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
zipp==3.20.2
